+++ src/Sylius/Component/Payment/phpunit.xml.dist
@@ -10,7 +10,7 @@
 
     <testsuites>
         <testsuite name="SyliusPaymentComponent Test Suite">
-            <directory>./Tests/</directory>
+            <directory>./tests/</directory>
         </testsuite>
     </testsuites>
 </phpunit>
