+++ src/Sylius/Component/Payment/tests/Model/PaymentTest.php
@@ -13,7 +13,6 @@ declare(strict_types=1);
 
 namespace Tests\Sylius\Component\Payment\Model;
 
-use DateTime;
 use PHPUnit\Framework\TestCase;
 use Sylius\Component\Payment\Model\Payment;
 use Sylius\Component\Payment\Model\PaymentInterface;
@@ -87,7 +86,7 @@ final class PaymentTest extends TestCase
 
     public function testItsCreationDateIsMutable(): void
     {
-        $date = new DateTime('last year');
+        $date = new \DateTime('last year');
 
         $this->payment->setCreatedAt($date);
         $this->assertSame($date, $this->payment->getCreatedAt());
@@ -100,7 +99,7 @@ final class PaymentTest extends TestCase
 
     public function testItsLastUpdateDateIsMutable(): void
     {
-        $date = new DateTime('last year');
+        $date = new \DateTime('last year');
 
         $this->payment->setUpdatedAt($date);
         $this->assertSame($date, $this->payment->getUpdatedAt());
